% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{manufacturer_model}
\alias{manufacturer_model}
\title{Manufacturer model}
\format{
\subsection{\code{manufacturer_model}}{

A brms model object in categorical family predicting a categorical response
variable \code{manufacturer} with \code{cyl} as predictors.
}
}
\usage{
manufacturer_model
}
\description{
An example Bayesian categorical model fitted on ggplot2::mpg dataset predicts
manufacturer for auotmobiles.
}
\keyword{datasets}
