% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_comparative_layout.R
\name{mc_comp_layout}
\alias{mc_comp_layout}
\title{Define comparative layout in model check visualization}
\usage{
mc_comp_layout(comparative_layout, ..., exp_op = NULL)
}
\arguments{
\item{comparative_layout}{Which comparative layout to use? If \code{"jux"}, \code{mcplot()}
puts geoms for data observations and geoms for model predictions side by side;
if \code{"sup"}, \code{mcplot()} overlays them; if \code{"exp"}, \code{mcplot()} will show comparison by
explicit-encoding depending on \code{exp_op}}

\item{...}{Other arguments passed on to \href{https://wilkelab.org/cowplot/articles/plot_grid.html}{\code{cowplot::plot_grid()}},
if using \code{"jux"} for comparative_layout.}

\item{exp_op}{What operation needs for explicit-encoding? If \code{exp_op} is:
\itemize{
\item \code{"residual"}, then \code{mcplot()} will show the comparison between data observations and
model predictions by a residual plot.
\item \code{"qq"}, then \code{mcplot()} will show the comparison by a Q-Q plot.
\item A function, then \code{mcplot()} will transform the data by \code{exp_op}.
}}
}
\description{
Define comparative layout in model check visualization
}
