% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_cond_var.R
\name{mc_cond_var}
\alias{mc_cond_var}
\title{Add marginal check to model check}
\usage{
mc_cond_var(x_var = NULL, row_vars = NULL, col_vars = NULL)
}
\arguments{
\item{x_var}{A variable quoted by \code{ggplot2::vars()} and defining the conditional variable on x axis.}

\item{row_vars, col_vars}{A set of variables quoted by vars() and defining faceting groups on the rows or columns dimension.}
}
\description{
Add marginal check to model check
}
