% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_cond_var.R
\name{mc_condition_on}
\alias{mc_condition_on}
\title{Add marginal check to model check}
\usage{
mc_condition_on(x = NULL, color = NULL, row = NULL, col = NULL)
}
\arguments{
\item{x}{A variable quoted by \code{ggplot2::vars()} and defining the conditional variable on x axis.}

\item{color}{A variable quoted by \code{ggplot2::vars()} and defining the color aesthetic in model check.}

\item{row, col}{A set of variables quoted by \code{ggplot2::vars()} and defining faceting groups on the rows or columns}
}
\description{
Add marginal check to model check
}
\examples{
library(ggplot2)
library(dplyr)

mcplot(mpg_model) +
  mc_condition_on(x = vars(disp))

mcplot(mpg_model) +
  mc_condition_on(x = vars(disp), row = vars(vs), col = vars(am))
}
