% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_distribution.R
\name{mc_distribution}
\alias{mc_distribution}
\title{Define how to draw from posterior distribution}
\usage{
mc_distribution(
  distribution = "predictive",
  newdata = "obs",
  is.transform = TRUE,
  ndraws = 500,
  seed = NULL,
  re_formula = NULL
)
}
\arguments{
\item{distribution}{Which distribution to draw from. The options include \code{"predictive"} and
and pushforward transformations (e.g. $\mu$, $\sigma$, and $\phi$). For example,
if the model is normal family and \code{distribution = "predictive"}, we draws from
posterior predictive distribution; if \code{distribution = "mu"}, we draws from
the linear/link-level predictor (i.e. pushforward transformations).}

\item{newdata}{Data frame to generate predictions from.}

\item{is.transform}{Should the linear predictor be transformed using the
inverse-link function? See \href{https://mc-stan.org/rstantools/reference/posterior_linpred.html}{\code{rstanarm::posterior_linpred()}} or
\href{https://mc-stan.org/rstantools/reference/posterior_linpred.html}{\code{brms::posterior_linpred()}}.}

\item{ndraws}{The number of draws to return, or \code{NULL} to return all draws.}

\item{seed}{A seed to use when subsampling draws (i.e. when ndraws is not NULL).}

\item{re_formula}{Formula containing group-level effects to be considered
in the prediction. See \href{http://mjskay.github.io/tidybayes/reference/add_predicted_draws.html}{\code{add_predicted_draws}}.}
}
\description{
Define how to draw from posterior distribution
}
