% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_gglayer.R
\name{mc_gglayer}
\alias{mc_gglayer}
\title{Add a \code{ggplot2::layer()} to model check visualization}
\usage{
mc_gglayer(layers)
}
\arguments{
\item{layers}{A ggplot layer or a list of ggplot layers.}
}
\description{
Since modelcheck uses \code{ggplot2} to implement the plots, you can use \code{mc_gglayer()}
to pass in a ggplot layer to the ggplot object generated by modelcheck.
}
