% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_uncertainty_representation.R
\name{mc_model}
\alias{mc_model}
\title{Define the geom/uncertainty representation for model}
\usage{
mc_model(uncertainty_representation, ...)
}
\arguments{
\item{uncertainty_representation}{The uncertainty representation option,
including ``}

\item{...}{Augments passed to geom function. For example, \code{mc_model("dots", size = 10)}
will pass \code{size = 10} to the geom function \href{https://mjskay.github.io/ggdist/reference/stat_dots.html}{\code{ggdist::stat_dots()}}.}
}
\description{
\code{mc_model()} adds geoms for draws from model posterior distribution, either
basic geom (e.g. points, lines, and tiles) or uncertainty representations
(e.g. eye plots, dots plots, and lineribbon plot). \code{mc_model_*(...)} are
equivalent to \code{mc_model("*", ...)}.
}
