% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_obs_plot.R,
%   R/mc_uncertainty_representation.R
\name{mc_obs_cdf}
\alias{mc_obs_cdf}
\alias{mc_model_cdf}
\title{CDF bar plot for model predictions}
\usage{
mc_obs_cdf(...)

mc_model_cdf(..., n_sample = NA, group_sample = "collapse")
}
\arguments{
\item{...}{Augments passed to \href{https://mjskay.github.io/ggdist/reference/stat_cdfinterval.html}{\code{ggdist::stat_cdfinterval}}.}

\item{n_sample}{The number of sample of draws to show in CDF bar plot.
Default \code{NA}. If \code{n_sample} is \code{NA}, then \code{mc_model_cdf()} will use all draws from
posterior distribution.}

\item{group_sample}{How to show samples? Default \code{"collapse"}. If \code{group_sample} is \code{"collapse"},
then all samples are collapsed together to show in one CDF bar plot; if \code{group_sample}
is \code{"group"}, then each sample is shown in an individual CDF bar plot; if
\code{group_sample} is \code{"hops"}, then \code{mc_model_cdf()} will use animation to show each
sample in one frame; if \code{group_sample} is an function, then all samples are aggregated
by \code{group_sample()}.}
}
\description{
CDF bar plot for model predictions
}
\examples{
library(ggplot2)

mcplot(mpg_model) +
  mc_model_cdf(n_sample = 50) +
  mc_obs_cdf() +
  mc_gglayer(coord_flip())

mcplot(mpg_model) +
  mc_model_cdf(n_sample = 50) +
  mc_obs_cdf() +
  mc_condition_on(x = vars(vs))

mcplot(mpg_model) +
  mc_model_cdf(n_sample = 50, group_sample = "group") +
  mc_obs_cdf() +
  mc_condition_on(x = vars(vs))

mcplot(mpg_model) +
  mc_model_cdf(n_sample = 50, group_sample = mean) +
  mc_observation_transformation(mean) +
  mc_obs_cdf() +
  mc_condition_on(x = vars(vs))
}
