% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_obs_plot.R,
%   R/mc_uncertainty_representation.R
\name{mc_obs_custom}
\alias{mc_obs_custom}
\alias{mc_model_custom}
\title{Customized geom}
\usage{
mc_obs_custom(plot, ...)

mc_model_custom(
  plot,
  ...,
  n_sample = NA,
  group_sample = "collapse",
  group_on = NULL
)
}
\arguments{
\item{plot}{The plot to use. It should be a geom function
like \href{https://ggplot2.tidyverse.org/reference/index.html#geoms}{geoms in \code{ggplots}}
and should be able to have \code{data} and \code{mapping} augments. See examples for more details.}

\item{...}{Augments passed to \code{plot}.}

\item{n_sample}{The number of sample of draws to show.
Default \code{NA}. If \code{n_sample} is \code{NA}, then \code{mc_model_custom()} will use all draws from
posterior distribution.}

\item{group_sample}{How to show draws? Default \code{"hops"}.
If \code{group_sample} is \code{"collapse"}, then all draws are collapsed together
and are shown in one geom; if \code{"draw"} is \code{"group"}, then only one draw
is shown in one geom;
if \code{group_sample} is \code{"hops"}, then \code{mc_model_custom()} will use animation to show each
draw in one frame; if \code{group_sample} is an function, then all draws are aggregated
by \code{group_sample()}. See examples for more details.}

\item{group_on}{To group the samples by sample id or row id (i.e., the input data point id).
Default \code{NULL}. If \code{group_on = NULL} or \code{group_on = "sample"}, the group sample method is applied by grouping on \code{.draw}.
If \code{group_on = "row"}, the group sample method is applied on \code{.row}.}
}
\description{
Customized geom
}
\examples{
library(ggplot2)
library(ggdist)

mcplot(mpg_model) +
  mc_model_custom(stat_boxplot, notch = TRUE) +
  mc_obs_custom(geom_swarm) +
  mc_condition_on(x = vars(vs)) +
  mc_gglayer(coord_flip())
}
