% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_obs_plot.R,
%   R/mc_uncertainty_representation.R
\name{mc_obs_line}
\alias{mc_obs_line}
\alias{mc_model_line}
\title{Line geom}
\usage{
mc_obs_line(...)

mc_model_line(..., n_sample = NA, group_sample = "group", group_on = NULL)
}
\arguments{
\item{...}{Augments passed to \href{https://ggplot2.tidyverse.org/reference/geom_path.html}{\code{ggplot2::geom_line}}.}

\item{n_sample}{The number of sample of draws to show.
Default \code{NA}. If \code{n_sample} is \code{NA}, then \code{mc_model_line()} will use all draws from
posterior distribution.}

\item{group_sample}{How to show draws? Default \code{"group"}.
If \code{group_sample} is \code{"collapse"}, then all draws are collapsed together
and are connected by one line; if \code{"draw"} is \code{"group"}, then data from each draw is
connected by one line;
if \code{group_sample} is \code{"hops"}, then \code{mc_model_line()} will use animation to show each
draw in one frame; if \code{group_sample} is an function, then all draws are aggregated
by \code{group_sample()}. See examples for more details.}

\item{group_on}{To group the samples by sample id or row id (i.e., the input data point id).
Default \code{NULL}. If \code{group_on = NULL} or \code{group_on = "sample"}, the group sample method is applied by grouping on \code{.draw}.
If \code{group_on = "row"}, the group sample method is applied on \code{.row}.}
}
\description{
Line geom
}
\examples{
library(ggplot2)

mcplot(mpg_model) +
  mc_model_line(n_sample = 50) +
  mc_obs_line() +
  mc_condition_on(x = vars(disp))
}
