% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_obs_plot.R,
%   R/mc_uncertainty_representation.R
\name{mc_obs_lineribbon}
\alias{mc_obs_lineribbon}
\alias{mc_model_lineribbon}
\title{Line + multiple-ribbon plot for model predictions}
\usage{
mc_obs_lineribbon(..., scale_fill = ggplot2::scale_fill_brewer(palette = 1))

mc_model_lineribbon(
  ...,
  scale_fill = ggplot2::scale_fill_brewer(palette = 7),
  n_sample = NA,
  draw = "collapse"
)
}
\arguments{
\item{...}{Augments passed to \href{https://mjskay.github.io/ggdist/reference/stat_lineribbon.html}{\code{ggdist::stat_lineribbon}}.}

\item{scale_fill}{The scale_fill_* layer for lineribbon. Default to be \code{ggplot2::scale_fill_brewer()}.}

\item{n_sample}{The number of sample of draws to show in lineribbon plot.
Default \code{NA}. If \code{n_sample} is \code{NA}, then \code{mc_model_lineribbon()} will use all draws from
posterior distribution.}

\item{draw}{How to show draws? Default \code{"collapse"}. If \code{draw} is \code{"collapse"},
then all draws are collapsed together to show in one lineribbon plot; if \code{draw}
is \code{"group"}, then each draw is shown in an individual lineribbon plot; if
\code{draw} is \code{"hops"}, then \code{mc_model_lineribbon()} will use animation to show each
draw in one frame; if \code{draw} is an function, then all draws are aggregated
by \code{draw()}. See examples for more details.}
}
\description{
Line + multiple-ribbon plot for model predictions
}
