% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_obs_plot.R,
%   R/mc_uncertainty_representation.R
\name{mc_obs_pointinterval}
\alias{mc_obs_pointinterval}
\alias{mc_model_pointinterval}
\title{Point + interval plot for model predictions}
\usage{
mc_obs_pointinterval(...)

mc_model_pointinterval(..., n_sample = NA, draw = "collapse")
}
\arguments{
\item{...}{Augments passed to \href{https://mjskay.github.io/ggdist/reference/stat_pointinterval.html}{\code{ggdist::stat_pointinterval}}.}

\item{n_sample}{The number of sample of draws to show in pointinterval plot.
Default \code{NA}. If \code{n_sample} is \code{NA}, then \code{mc_model_pointinterval()} will use all draws from
posterior distribution.}

\item{draw}{How to show draws? Default \code{"collapse"}. If \code{draw} is \code{"collapse"},
then all draws are collapsed together to show in one pointinterval plot; if \code{draw}
is \code{"group"}, then each draw is shown in an individual pointinterval plot; if
\code{draw} is \code{"hops"}, then \code{mc_model_pointinterval()} will use animation to show each
draw in one frame; if \code{draw} is an function, then all draws are aggregated
by \code{draw()}. See examples for more details.}
}
\description{
Point + interval plot for model predictions
}
\examples{
library(ggplot2)

mcplot(mpg_model) +
  mc_model_pointinterval(n_sample = 50) +
  mc_obs_pointinterval() +
  mc_gglayer(coord_flip())

mcplot(mpg_model) +
  mc_model_pointinterval(n_sample = 50) +
  mc_obs_pointinterval() +
  mc_condition_on(x = vars(disp))
}
