% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_obs_plot.R,
%   R/mc_uncertainty_representation.R
\name{mc_obs_reference_line}
\alias{mc_obs_reference_line}
\alias{mc_model_reference_line}
\title{Horizontal reference lines on y axis}
\usage{
mc_obs_reference_line(...)

mc_model_reference_line(..., n_sample = NA, group_sample = "collapse")
}
\arguments{
\item{...}{Augments passed to \href{https://ggplot2.tidyverse.org/reference/geom_abline.html}{\code{ggplot2::geom_hline}}.}

\item{n_sample}{The number of sample of draws to show.
Default \code{NA}. If \code{n_sample} is \code{NA}, then \code{mc_model_tile()} will use all draws from
posterior distribution.}

\item{group_sample}{How to show draws? Default \code{"collapse"}.
If \code{group_sample} is \code{"collapse"}, then all draws are collapsed together
and are shown in one reference line; if \code{"draw"} is \code{"group"}, then only one draw
is shown by reference line;
if \code{group_sample} is \code{"hops"}, then \code{mc_model_reference_line()} will use animation to show each
draw in one frame; if \code{group_sample} is an function, then all draws are aggregated
by \code{group_sample()}. See examples for more details.}
}
\description{
Horizontal reference lines on y axis
}
\examples{
library(ggplot2)

mcplot(mpg_model) +
  mc_observation_transformation(mean) +
  mc_model_slab(n_sample = 50) +
  mc_obs_reference_line()
}
