% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_obs_plot.R
\name{mc_obs}
\alias{mc_obs}
\title{Define the geom/uncertainty representation for data}
\usage{
mc_obs(obs_uncertainty_representation, ...)
}
\arguments{
\item{...}{Augments passed to geom function. For example, \code{mc_obs("dots", size = 10)}
will pass \code{size = 10} to the geom function \href{https://mjskay.github.io/ggdist/reference/stat_dots.html}{\code{ggdist::stat_dots()}}.}

\item{uncertainty_representation}{The uncertainty representation option,
including ``}
}
\description{
\code{mc_obs()} adds geoms for data observations, either
basic geom (e.g. points, lines, and tiles) or uncertainty representations
(e.g. eye plots, dots plots, and lineribbon plot). \code{mc_obs_*(...)} are
equivalent to \code{mc_obs("*", ...)}.
}
