% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_observation_transformation.R
\name{mc_observation_transformation}
\alias{mc_observation_transformation}
\title{Define the transformation applied to observed data}
\usage{
mc_observation_transformation(transform = NULL, group = NULL)
}
\arguments{
\item{transform}{The transform function that is applied on the
response variable in observed data. Default to be \code{NULL}. If \code{NULL},
\code{vmc} will use no transformation. The transform function takes an input
of a data frame, e.g., the data frame passed to \code{observation}, containing a
column named \code{observation} standing for observations of the response variable
and several columns standing for the predictors in model (if any). The output
of the transform function should be in the same form as the input, a data frame
containing a column named observation and several columns for the predictors.
This argument is useful when \code{mc_distribution} is set to a distribution
that is in a different unit from the raw observation, e.g., \code{sigma} in Gaussian
family models describes the variance of observation. See example for more details.}

\item{group}{A set of variables quoted by \code{ggplot2::vars()} and represents
the group option used before applying the \code{transform} function.}
}
\description{
\code{mc_observation_transformation()} defines the necessary transformation applied
to the observed data before passing it to generate model check.
}
\details{
The transformation is necessary when the distribution generated by the model
is not predictive distributions. For example, when the distribution generated
from \code{phi} parameter of beta distribution, the observed data can not be shown
in model check directly but need to be transformed into the same scale of
\code{phi}.
}
\examples{
library(ggplot2)

mcplot(mpg_model) +
  mc_distribution("sigma") +
  mc_observation_transformation(sd, vars(disp)) +
  mc_gglayer(coord_flip())
}
