% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_plot.R
\name{mcplot}
\alias{mcplot}
\title{Create a new mcplot}
\usage{
mcplot(model, observation = NULL)
}
\arguments{
\item{model}{The model fit object.}

\item{observation}{A data frame standing for data observations. Default to be
\code{NULL}. If \code{NULL}, \code{modelcheck} will use the observations in the model fit
data set. The input data frame should include the variables in model formula.}
}
\description{
\code{mcplot()} initializes a modelcheck object. It follows the ggplot syntax,
which uses a plus sign (\code{+}) to define features of model check visualization.
}
\details{
\code{mcplot()} uses a list of defaults to generate model check visualizations.
One line of \code{mcplot(model)} could generate a complete visualization for
posterior predictive checks. See \code{vignette("modelcheck")} for a complete
guidance.
}
