% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelcheck-package.R
\docType{package}
\name{modelcheck-package}
\alias{modelcheck}
\alias{modelcheck-package}
\title{modelcheck: Check your model with data observations}
\description{
Check your model with data observations
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mucollective.github.io/modelcheck/}
}

}
\author{
\strong{Maintainer}: Ziyang Guo \email{ziyangguo2027@u.northwestern.edu}

Authors:
\itemize{
  \item Jessica Hullman \email{jessica.hullman@gmail.com}
  \item Matthew Kay \email{mjskay@northwestern.edu}
  \item Alex Kale \email{kalea@uchicago.edu}
}

}
\keyword{internal}
