% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mpg_model}
\alias{mpg_model}
\title{Miles/(US) gallon model}
\format{
An object of class \code{brmsfit} of length 22.
}
\usage{
mpg_model
}
\description{
An example Bayesian gaussian model fitted on mtcars dataset predicts
miles/(US) gallon for auotmobiles.
}
\keyword{datasets}
