% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_distribution.R
\name{mc_distribution}
\alias{mc_distribution}
\title{Define how to draw from posterior distribution}
\usage{
mc_distribution(
  distribution = "prediction",
  newdata = NULL,
  extract_function = NULL,
  response_var = NULL,
  ndraws = 500,
  transform = TRUE,
  ...
)
}
\arguments{
\item{distribution}{Which distribution to draw from. The options include \code{"prediction"} and
and push-forward transformations (e.g. \code{mu}, \code{sigma}, and \code{phi}). For example,
if the model is normal family and \code{distribution = "prediction"}, we draws from
posterior predictive distribution; if \code{distribution = "mu"}, we draws from
the distribution of linear/link-level predictor (i.e. push-forward transformations).}

\item{newdata}{Data frame to generate predictions from, or \code{NULL} to reuse the
data frame used to fit model, i.e. replicated predictive distribution.}

\item{extract_function}{The function used to draw from model's posterior distributions.
Default to be \code{NULL}. If \code{extract_function} is \code{NULL}, then \code{vmc} will
use \code{tidybayes::predicted_draws()} for \code{distribution = "prediction"} and
use \code{tidybayes::linpred_draws()} for other distributions of
linear/link-level predictors. If \code{extract_function} is not \code{NULL}, it should be
a function that takes \code{model}, \code{newdata}, and \code{ndraws} as inputs and output a data
frame that sampled from \code{model} on \code{newdata} with at least a \code{prediction}
column (the draws from the model's distribution), a \code{.row} column (a factor grouping rows from
the input newdata), and a \code{.draw} column (a unique index corresponding to
each draw from the distribution).}

\item{response_var}{A string for the response variable in model. Default to
be \code{NULL}. If \code{NULL}, \code{vmc} will infer the response variable from
\code{model$formula}.}

\item{ndraws}{The number of draws to return, or \code{NULL} to return all draws.}

\item{...}{Augments passed to \code{extract_function}. If \code{extract_function} is \code{NULL},
then \code{...} will be passed to \code{tidybayes::predicted_draws()} or \code{tidybayes::linpred_draws()}}
}
\description{
\code{vmc} uses the R package \href{http://mjskay.github.io/tidybayes/index.html}{\code{tidybayes}}
to draw from posterior distributions. For posterior predictive distributions,
\code{vmc} uses \code{tidybayes::predicted_draws()}; for posterior distribution
of push-forward transformation, \code{vmc} uses \code{tidybayes::linpred_draws()}.
}
\examples{
library(tidybayes)
library(dplyr)
library(modelr)

mcplot(mpg_model) +
  mc_distribution()

mcplot(mpg_model) +
  mc_distribution("mu")

mcplot(mpg_model) +
  mc_distribution("mu", data_grid(mpg_model$data, disp, vs, am))

epred_draws_mu = function(model, newdata, ...) {
  epred_draws(model, newdata, dpar = "mu", ...) \%>\%
    mutate(prediction = mu)
}

mcplot(mpg_model) +
  mc_distribution(extract_function = epred_draws_mu)
}
