% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_gglayer.R
\name{mc_gglayer}
\alias{mc_gglayer}
\title{Pass \code{ggplot2::layer()} to model check visualization}
\usage{
mc_gglayer(layers)
}
\arguments{
\item{layers}{A list of \code{ggplot2::layer()}.}
}
\description{
Pass \code{ggplot2::layer()} to model check visualization
}
