% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_uncertainty_representation.R
\name{mc_model}
\alias{mc_model}
\alias{mc_model_cdf}
\alias{mc_model_ccdf}
\alias{mc_model_dots}
\alias{mc_model_dotsinterval}
\alias{mc_model_eye}
\alias{mc_model_halfeye}
\alias{mc_model_slab}
\alias{mc_model_gradient}
\alias{mc_model_histinterval}
\alias{mc_model_pointinterval}
\alias{mc_model_interval}
\alias{mc_model_lineribbon}
\alias{mc_model_ribbon}
\alias{mc_model_geom}
\alias{mc_model_geom_auto}
\alias{mc_model_geom_point}
\alias{mc_model_geom_line}
\alias{mc_model_geom_tile}
\alias{mc_model_geom_custom}
\title{Define the geom/uncertainty representation for model}
\usage{
mc_model(uncertainty_representation, ...)

mc_model_cdf(...)

mc_model_ccdf(...)

mc_model_dots(...)

mc_model_dotsinterval(...)

mc_model_eye(...)

mc_model_halfeye(...)

mc_model_slab(...)

mc_model_gradient(...)

mc_model_histinterval(...)

mc_model_pointinterval(...)

mc_model_interval(...)

mc_model_lineribbon(...)

mc_model_ribbon(...)

mc_model_geom(..., base_plot = "auto")

mc_model_geom_auto(...)

mc_model_geom_point(...)

mc_model_geom_line(...)

mc_model_geom_tile(...)

mc_model_geom_custom(plot, ...)
}
\arguments{
\item{uncertainty_representation}{The uncertainty representation option,
including ``}

\item{...}{Augments passed to geom function. For example, \code{mc_model("dots", size = 10)}
will pass \code{size = 10} to the geom function \href{https://mjskay.github.io/ggdist/reference/stat_dots.html}{\code{stat_dots()}}.}
}
\description{
\code{mc_model()} adds geoms for draws from model posterior distribution, either
basic geom (e.g. points, lines, and tiles) or uncertainty representations
(e.g. eye plots, dots plots, and lineribbon plot). \code{mc_model_*(...)} are
equivalent to \code{mc_model("*", ...)}.
}
