% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_obs_plot.R,
%   R/mc_uncertainty_representation.R
\name{mc_obs_custom}
\alias{mc_obs_custom}
\alias{mc_model_custom}
\title{Customized geom}
\usage{
mc_obs_custom(plot, ...)

mc_model_custom(geom_plot, ..., n_sample = NA, draw = "collapse")
}
\arguments{
\item{...}{Augments passed to \code{geom_plot}.}

\item{geom_plot}{The geom plot to use. \code{"geom_plot"} should be a geom function
like \href{https://ggplot2.tidyverse.org/reference/index.html#geoms}{geoms in \code{ggplots}}
and should be able to have \code{data} and \code{mapping} augments. See examples for more details.}

\item{n_sample}{The number of sample of draws to show.
Default \code{NA}. If \code{n_sample} is \code{NA}, then \code{mc_model_custom()} will use all draws from
posterior distribution.}

\item{draw}{How to show draws? Default \code{"hops"}.
If \code{draw} is \code{"collapse"}, then all draws are collapsed together
and are shown in one geom; if \code{"draw"} is \code{"group"}, then only one draw
is shown in one geom;
if \code{draw} is \code{"hops"}, then \code{mc_model_custom()} will use animation to show each
draw in one frame; if \code{draw} is an function, then all draws are aggregated
by \code{draw()}. See examples for more details.}
}
\description{
Customized geom
}
