% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_uncertainty_representation.R
\name{mc_model_eye}
\alias{mc_model_eye}
\title{Eye (violin + interval) plot for model predictions}
\usage{
mc_model_eye(..., n_sample = NA, draw = "collapse")
}
\arguments{
\item{...}{Augments passed to \href{https://mjskay.github.io/ggdist/reference/stat_eye.html}{\code{ggdist::stat_eye}}.}

\item{n_sample}{The number of sample of draws to show in eye plot.
Default \code{NA}. If \code{n_sample} is \code{NA}, then \code{mc_model_eye()} will use all draws from
posterior distribution.}

\item{draw}{How to show draws? Default \code{"collapse"}. If \code{draw} is \code{"collapse"},
then all draws are collapsed together to show in one eye plot; if \code{draw}
is \code{"group"}, then each draw is shown in an individual eye plot; if
\code{draw} is \code{"hops"}, then \code{mc_model_eye()} will use animation to show each
draw in one frame; if \code{draw} is an function, then all draws are aggregated
by \code{draw()}. See examples for more details.}
}
\description{
Eye (violin + interval) plot for model predictions
}
