% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_uncertainty_representation.R
\name{mc_model_gradientinterval}
\alias{mc_model_gradientinterval}
\title{Gradient + interval plot for model predictions}
\usage{
mc_model_gradientinterval(..., n_sample = NA, draw = "collapse")
}
\arguments{
\item{...}{Augments passed to \href{https://mjskay.github.io/ggdist/reference/stat_gradientinterval.html}{\code{ggdist::stat_gradientinterval}}.}

\item{n_sample}{The number of sample of draws to show in gradient + interval plot.
Default \code{NA}. If \code{n_sample} is \code{NA}, then \code{mc_model_gradientinterval()} will use all draws from
posterior distribution.}

\item{draw}{How to show draws? Default \code{"collapse"}. If \code{draw} is \code{"collapse"},
then all draws are collapsed together to show in one gradient + interval plot; if \code{draw}
is \code{"group"}, then each draw is shown in an individual gradient + interval plot; if
\code{draw} is \code{"hops"}, then \code{mc_model_gradientinterval()} will use animation to show each
draw in one frame; if \code{draw} is an function, then all draws are aggregated
by \code{draw()}. See examples for more details.}
}
\description{
Gradient + interval plot for model predictions
}
