% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_obs_plot.R,
%   R/mc_uncertainty_representation.R
\name{mc_obs_halfeye}
\alias{mc_obs_halfeye}
\alias{mc_model_halfeye}
\title{Half-eye (density + interval) plot for model predictions}
\usage{
mc_obs_halfeye(...)

mc_model_halfeye(
  ...,
  n_sample = NA,
  group_sample = "collapse",
  group_on = NULL
)
}
\arguments{
\item{...}{Augments passed to \href{https://mjskay.github.io/ggdist/reference/stat_halfeye.html}{\code{ggdist::stat_halfeye}}.}

\item{n_sample}{The number of sample of draws to show in half eye plot.
Default \code{NA}. If \code{n_sample} is \code{NA}, then \code{mc_model_halfeye()} will use all draws from
posterior distribution.}

\item{group_sample}{How to show draws? Default \code{"collapse"}. If \code{group_sample} is \code{"collapse"},
then all draws are collapsed together to show in one half eye plot; if \code{group_sample}
is \code{"group"}, then each draw is shown in an individual half eye plot; if
\code{group_sample} is \code{"hops"}, then \code{mc_model_halfeye()} will use animation to show each
draw in one frame; if \code{group_sample} is an function, then all draws are aggregated
by \code{group_sample()}. See examples for more details.}

\item{group_on}{To group the samples by sample id or row id (i.e., the input data point id).
Default \code{NULL}. If \code{group_on = NULL} or \code{group_on = "sample"}, the group sample method is applied by grouping on \code{.draw}.
If \code{group_on = "row"}, the group sample method is applied on \code{.row}.}
}
\description{
Half-eye (density + interval) plot for model predictions
}
\examples{
library(ggplot2)

mcplot(mpg_model) +
  mc_model_halfeye(n_sample = 50) +
  mc_obs_halfeye() +
  mc_gglayer(coord_flip())

mcplot(mpg_model) +
  mc_model_halfeye(n_sample = 50) +
  mc_obs_halfeye() +
  mc_condition_on(x = vars(vs))

mcplot(mpg_model) +
  mc_model_halfeye(n_sample = 50, group_sample = mean) +
  mc_obs_halfeye() +
  mc_condition_on(x = vars(vs))

mcplot(mpg_model) +
  mc_model_halfeye(n_sample = 50, group_sample = mean, group_on = "row") +
  mc_obs_halfeye() +
  mc_condition_on(x = vars(vs))
}
