% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_obs_plot.R,
%   R/mc_uncertainty_representation.R
\name{mc_obs_interval}
\alias{mc_obs_interval}
\alias{mc_model_interval}
\title{Interval plot for model predictions}
\usage{
mc_obs_interval(..., scale_color = ggplot2::scale_color_brewer(palette = 1))

mc_model_interval(
  ...,
  scale_color = ggplot2::scale_color_brewer(palette = 7),
  n_sample = NA,
  draw = "collapse"
)
}
\arguments{
\item{...}{Augments passed to \href{https://mjskay.github.io/ggdist/reference/stat_interval.html}{\code{ggdist::stat_interval}}.}

\item{scale_color}{The scale_color_* layer for interval. Default to be \code{ggplot2::scale_color_brewer()}.}

\item{n_sample}{The number of sample of draws to show in interval plot.
Default \code{NA}. If \code{n_sample} is \code{NA}, then \code{mc_model_interval()} will use all draws from
posterior distribution.}

\item{draw}{How to show draws? Default \code{"collapse"}. If \code{draw} is \code{"collapse"},
then all draws are collapsed together to show in one interval plot; if \code{draw}
is \code{"group"}, then each draw is shown in an individual interval plot; if
\code{draw} is \code{"hops"}, then \code{mc_model_interval()} will use animation to show each
draw in one frame; if \code{draw} is an function, then all draws are aggregated
by \code{draw()}. See examples for more details.}
}
\description{
Interval plot for model predictions
}
