% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_obs_plot.R,
%   R/mc_uncertainty_representation.R
\name{mc_obs_point}
\alias{mc_obs_point}
\alias{mc_model_point}
\title{Points geom}
\usage{
mc_obs_point(...)

mc_model_point(..., n_sample = NA, draw = "collapse")
}
\arguments{
\item{...}{Augments passed to \href{https://ggplot2.tidyverse.org/reference/geom_point.html}{\code{ggplot2::geom_point}}.}

\item{n_sample}{The number of sample of draws to show.
Default \code{NA}. If \code{n_sample} is \code{NA}, then \code{mc_model_point()} will use all draws from
posterior distribution.}

\item{draw}{How to show draws? Default \code{"collapse"}. \verb{"collapse} and \verb{"group}
are same for point geom, since point geom is just showing each data points.
If \code{draw} is \code{"collapse"} or \code{"group"},
then all draws are collapsed together and each point represents one data point;
if \code{draw} is \code{"hops"}, then \code{mc_model_point()} will use animation to show each
draw in one frame; if \code{draw} is an function, then all draws are aggregated
by \code{draw()}. See examples for more details.}
}
\description{
Points geom
}
