% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_obs_plot.R,
%   R/mc_uncertainty_representation.R
\name{mc_obs_tile}
\alias{mc_obs_tile}
\alias{mc_model_tile}
\title{Tile geom}
\usage{
mc_obs_tile(...)

mc_model_tile(..., n_sample = NA, draw = "hops")
}
\arguments{
\item{...}{Augments passed to \href{https://ggplot2.tidyverse.org/reference/geom_tile.html}{\code{ggplot2::geom_tile}}.}

\item{n_sample}{The number of sample of draws to show.
Default \code{NA}. If \code{n_sample} is \code{NA}, then \code{mc_model_tile()} will use all draws from
posterior distribution.}

\item{draw}{How to show draws? Default \code{"hops"}.
If \code{draw} is \code{"collapse"}, then all draws are collapsed together
and are shown in one tile plot; if \code{"draw"} is \code{"group"}, then only one draw
is shown in tile plot;
if \code{draw} is \code{"hops"}, then \code{mc_model_tile()} will use animation to show each
draw in one frame; if \code{draw} is an function, then all draws are aggregated
by \code{draw()}. See examples for more details.}
}
\description{
Tile geom
}
