% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_set_order.R
\name{mc_set_order}
\alias{mc_set_order}
\title{Order ticks on axis, rows, and columns}
\usage{
mc_set_order(x = NULL, y = NULL, row = NULL, col = NULL)
}
\arguments{
\item{x, y, row, col}{A list of the order of x ticks, y ticks, rows, and columns.}
}
\description{
\code{mc_set_order} orders ticks on axis or grids on rows and columns. To make it ordered,
the variables on axis, rows, and columns must be discrete.
}
