% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_plot.R
\name{mcplot}
\alias{mcplot}
\title{Create a new mcplot}
\usage{
mcplot(model, observation = NULL)
}
\arguments{
\item{model}{The model fit object.}

\item{observation}{A data frame standing for data observations. Default to be
\code{NULL}. If \code{NULL}, \code{vmc} will set the observations as the
data set that used to fit the model. The input data frame should include the variables in model formula.}
}
\description{
\code{mcplot()} initializes a \code{vmc} object. It follows the ggplot syntax,
which uses a plus sign (\code{+}) to define features of model check visualization.
}
\details{
\code{mcplot()} uses a list of defaults to generate model check visualizations.
One line of \code{mcplot(model)} could generate a complete visualization for
posterior predictive checks. See \code{vignette("vmc")} for a complete
guidance.
}
\examples{
library(ggplot2)
library(dplyr)

mcplot(mpg_model)
# note the density is on x axis and the response variable, mpg, is on y axis.
# But you can flip the coordinates by mc_gglayer()
mcplot(mpg_model) +
  mc_gglayer(coord_flip())
# you can also choose to use another observed data to show in model checks
new_observed_data = mtcars \%>\% mutate(mpg = rnorm(nrow(mtcars), 20, 5))
mcplot(mpg_model, new_observed_data) +
  mc_gglayer(coord_flip())
}
