% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_plot.R
\name{mcplot}
\alias{mcplot}
\title{Create a new mcplot}
\usage{
mcplot(model)
}
\arguments{
\item{model}{The model fit object.}
}
\description{
\code{mcplot()} initializes a modelcheck object. It's
almost always followed by a plus sign (\code{+}) to add
components to the plot.
}
