% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-construction.R
\name{+.modelcheck}
\alias{+.modelcheck}
\title{Add components to a plot}
\usage{
\method{+}{modelcheck}(e1, e2)
}
\arguments{
\item{e1}{An object of class \code{modelcheck}.}

\item{e2}{A model check component, as described below.}
}
\description{
Similar to \code{ggplot2}, \code{vmc} uses \code{+} to allows you to start simple,
then get more and more complex, checking your work at each step.
}
\section{what can you add?}{

You can add any of the following types of objects:
\itemize{
\item An \code{mc_draw()} object overrides the current model distribution
\item A uncertainty representation layer created by \code{mc_model_} or \code{mc_obs_}
functions adds a visual layer for model distribution or observed data.
\item A comparative layout layer created by \code{mc_layout_} functions overrides
the current comparative layout.
\item A \code{mc_condition_on()} object overrides the current conditional variables
on x axis, rows, or columns.
\item A \code{mc_gglayer()} object adds a ggplot layer to the ggplot object
generated by \code{vmc}.
\item A color object overrides the color/fill of model distribution and
observed data in model check.
}

You can also supply a list, in which case each element of the list will
be added in turn.
}

\examples{
library(ggplot2)

mcplot(mpg_model) +
  mc_draw("mu") +
  mc_condition_on(x = vars(disp))

# Alternatively, you can add multiple components with a list.
# This can be useful to return from a function.
mcplot(mpg_model) +
  list(mc_draw("mu"),
       mc_condition_on(x = vars(disp)))
}
