% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mpg_model}
\alias{mpg_model}
\title{Miles/(US) gallon model}
\format{
\subsection{\code{mpg_model}}{

A brms model object in Gaussian family predicting a quantitative response
variable \code{mpg} with \code{disp}, \code{vs}, \code{am} as predictors.
}
}
\usage{
mpg_model
}
\description{
An example Bayesian gaussian model fitted on mtcars dataset predicts
miles/(US) gallon for auotmobiles.
}
\keyword{datasets}
